Migrate verifier and paypro files
