+++ tsconfig.json
@@ -4,7 +4,13 @@
     "module": "commonjs",
     "allowJs": true,
     "outDir": "dist",
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    "typeRoots": [
+      "./node_modules/@types"
+    ],
+    "types": [
+      "node"
+    ]
   },
   "include": [
     "lib/**/*"
